Ensemblers CRUD API (#55)
* - draft openspec api for batch ensembling
- refactor REST controller
- refactor ID usage of the model types

* - fix linter errors

* - pyfunc ensembler db migration
- ensemblers service + integration test

* - fix tests

* - List ensemblers API

* - Fix incorrect integration tests for router_service and router_version_service
- GetEnsembler API endpoint
- Updated docker-compose to switch between API v2 and batch

* - API method for saving new ensemblers
- Integration test to test the unmarshalling of the payload into one of ensembler concrete structs
- Updates to openapi spec

* - fix typo

* - refactor controllers to make them compatible with schema decoder and validation of query params

* - update ensembler endpoint

* - addressing PR comments:
  - added comments
  - made controller receiver style consistent
  - fixed down db migration
  - removed redundand Preloading in RouterService.Delete method

* - use named struct for path options

+++ api/turing/service/router_service.go
@@ -9,16 +9,21 @@ import (
 
 // RoutersService is the data access object for the Routers from the db.
 type RoutersService interface {
-	// List routers within the given project and environment.
-	ListRouters(projectID int, environmentName string) ([]*models.Router, error)
+	//ListRouters List routers within the given project and environment.
+	ListRouters(projectID models.ID, environmentName string) ([]*models.Router, error)
 	// Save the given router to the db. Updates the existing record if already exists.
 	Save(router *models.Router) (*models.Router, error)
-	// Find a router matching the given router id.
-	FindByID(routerID uint) (*models.Router, error)
-	// Find a router within the given project that matches the given name.
-	FindByProjectAndName(projectID int, routerName string) (*models.Router, error)
-	// Find a router within the given project and environment that matches the given name.
-	FindByProjectAndEnvironmentAndName(projectID int, environmentName string, routerName string) (*models.Router, error)
+	//FindByID Find a router matching the given router id.
+	FindByID(routerID models.ID) (*models.Router, error)
+	//FindByProjectAndName Find a router within the given project that matches the given name.
+	FindByProjectAndName(projectID models.ID, routerName string) (*models.Router, error)
+	//FindByProjectAndEnvironmentAndName Find a router within the given project and environment
+	// that matches the given name.
+	FindByProjectAndEnvironmentAndName(
+		projectID models.ID,
+		environmentName string,
+		routerName string,
+	) (*models.Router, error)
 	// Delete a router. This deletes all child objects of the router (router versions, ensemblers and enrichers)
 	// (Transactional).
 	Delete(router *models.Router) error
@@ -37,11 +42,10 @@ func (service *routersService) query() *gorm.DB {
 	return service.db.
 		Preload("CurrRouterVersion").
 		Preload("CurrRouterVersion.Enricher").
-		Preload("CurrRouterVersion.Ensembler").
-		Select("routers.*")
+		Preload("CurrRouterVersion.Ensembler")
 }
 
-func (service *routersService) ListRouters(projectID int, environmentName string) ([]*models.Router, error) {
+func (service *routersService) ListRouters(projectID models.ID, environmentName string) ([]*models.Router, error) {
 	var routers []*models.Router
 	query := service.query()
 	if projectID > 0 {
@@ -67,7 +71,7 @@ func (service *routersService) Save(router *models.Router) (*models.Router, erro
 	return service.FindByID(router.ID)
 }
 
-func (service *routersService) FindByID(routerID uint) (*models.Router, error) {
+func (service *routersService) FindByID(routerID models.ID) (*models.Router, error) {
 	var router models.Router
 	query := service.query().
 		Where("routers.id = ?", routerID).
@@ -79,7 +83,7 @@ func (service *routersService) FindByID(routerID uint) (*models.Router, error) {
 }
 
 func (service *routersService) FindByProjectAndEnvironmentAndName(
-	projectID int,
+	projectID models.ID,
 	environmentName string,
 	name string,
 ) (*models.Router, error) {
@@ -95,7 +99,7 @@ func (service *routersService) FindByProjectAndEnvironmentAndName(
 	return &router, nil
 }
 
-func (service *routersService) FindByProjectAndName(projectID int, name string) (*models.Router, error) {
+func (service *routersService) FindByProjectAndName(projectID models.ID, name string) (*models.Router, error) {
 	var router models.Router
 	query := service.query().
 		Where("routers.project_id = ?", projectID).
@@ -112,21 +116,26 @@ func (service *routersService) Delete(router *models.Router) error {
 		return errors.New("router must have valid primary key to be deleted")
 	}
 	tx := service.db.Begin()
-	var routerVersions []*models.RouterVersion
 
 	// remove associations
 	router.ClearCurrRouterVersion()
 	tx.Save(router)
 
-	service.db.Model(router).Related(&routerVersions)
+	var routerVersions []*models.RouterVersion
+	if err := service.db.
+		Where("router_id = ?", router.ID).
+		Find(&routerVersions).Error; err != nil {
+		return err
+	}
+
 	for _, routerVersion := range routerVersions {
-		if routerVersion.Ensembler != nil {
-			tx.Delete(routerVersion.Ensembler)
-		}
-		if routerVersion.Enricher != nil {
-			tx.Delete(routerVersion.Enricher)
-		}
 		tx.Delete(routerVersion)
+		if routerVersion.EnricherID.Valid {
+			tx.Delete(models.Enricher{}, "id = ?", routerVersion.EnricherID.Int32)
+		}
+		if routerVersion.EnsemblerID.Valid {
+			tx.Delete(models.Ensembler{}, "id = ?", routerVersion.EnsemblerID.Int32)
+		}
 	}
 	tx.Delete(router)
